pyproject.toml
[tool.poetry]
name = "crypto-quant-bot"
version = "0.1.0"
description = "Async crypto trading bot with Kalman Filter execution"
authors = ["Tim Peng <tim.e.peng@outlook.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ccxt = "^4.0"           # Crypto Exchange API
pandas = "^2.0"
numpy = "^1.24"
scikit-learn = "^1.2"
scipy = "^1.10"         # For Kalman Filter math
questdb = "^1.0"        # High-performance time-series DB (optional)
# OR use aiohttp if you write custom collectors:
aiohttp = "^3.8"
hyperliquid-python-sdk = "^0.21.0"
fastapi = "^0.124.0"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
torch = "^2.2"
statsmodels = "^0.14"
numba = "^0.58"
pyarrow = "^22.0.0"
vectorbt = "^0.28.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
pytest-asyncio = "^0.21" # Needed to test async functions
black = "^23.0"
jupyter = "^1.0"
matplotlib = "^3.10.7"
seaborn = "^0.13"
plotly = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"